Crash on a plaintext client request.
I discovered when running an FS2 TCP with TLS on a publicly open port , and it seems somebody was scanning my ports.

The whole application crashes and it is well before we get to doing anything with the socket.

The crash message is `javax.net.ssl.SSLException: Unrecognized SSL message, plaintext connection?`.

In particular, it seems no errors are handled at `fs2.io.net.tls.TLSEngine$$anon$1.$anonfun$unwrapHandshake$1(TLSEngine.scala:242)`, and I could not find any info about error handling with TLS in the docs.